% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arith_add.R
\name{\%add\%}
\alias{\%add\%}
\title{Add aribtrarily long integers}
\usage{
a \%add\% b
}
\arguments{
\item{a}{lhs. A big integer as a character string.}

\item{b}{rhs. A big integer as a character string.}
}
\value{
a + b (as a bit integer in a character string)
}
\description{
Split each number into chunks of add digits. Add these chunks in sequence
using R's built in interger addition (as the result is still within
R's 32bit integer type).
}
\details{
Manually handle the carry from this addition.
}
