% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arith_mul.R
\name{\%div\%}
\alias{\%div\%}
\title{Integer division of 2 long integers}
\usage{
a \%div\% b
}
\arguments{
\item{a}{lhs. A big integer as a character string.}

\item{b}{rhs. A big integer as a character string.}
}
\value{
a %/% b (as a bit integer in a character string)
}
\description{
This is quite slow as it uses the technique of 'long division'
\url{https://en.wikipedia.org/wiki/Division_algorithm}
}
