% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{\%gte\%}
\alias{\%gte\%}
\title{Greater than for TAPIR}
\usage{
lhs \%gte\% rhs
}
\arguments{
\item{lhs}{A big integer as a character string.}

\item{rhs}{A big integer as a character string.}
}
\value{
boolean
}
\description{
Greater than for TAPIR
}
