% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{\%lt\%}
\alias{\%lt\%}
\title{Less than for TAPIR}
\usage{
lhs \%lt\% rhs
}
\arguments{
\item{lhs}{A big integer as a character string.}

\item{rhs}{A big integer as a character string.}
}
\value{
boolean
}
\description{
Less than for TAPIR
}
