% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbigint.R
\name{rbigint}
\alias{rbigint}
\title{Generate a an arbitrary integer as a text string}
\usage{
rbigint(digits)
}
\arguments{
\item{digits}{integer. number of digits}
}
\value{
A character string containing an integer with the specified number of digits.
}
\description{
Generate a an arbitrary integer as a text string
}
