% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_into_chunks.R
\name{split_into_chunks}
\alias{split_into_chunks}
\title{Split numbers A and B into chunks of N digits}
\usage{
split_into_chunks(a, b, N)
}
\arguments{
\item{a}{first number}

\item{b}{second number}

\item{N}{number of digits in chunk}
}
\description{
Split numbers A and B into chunks of N digits
}
