\name{Topics0105}
\alias{Topics0105}
\docType{data}
\title{ Topics of House of Commons Divisions, 2001-2005 }
\description{
  A data frame containing one variable, the Hansard textual tag
  for the topic of each division.  Dvision identifiers are row names.
}
\usage{data(Topics0105)}
\format{
  A data frame with 1246 observations on the following variable.
  \describe{
    \item{\code{topic}}{a character vector}
     }
}
\references{
  Firth, D and Spirling, A (2003) Divisions of the United Kingdom House of 
  Commons, from 1992 to 2003 and beyond.  Submitted for publication.  At
  \url{http://www.warwick.ac.uk/go/tapir}.
}
\seealso{\code{\link{Votes0105}}}
\examples{
data(Topics0105)
}
\keyword{datasets}
